New changes
This version is based on the solution downloaded from the master file at 7pm November 10. things changed in this version:

1. Deleted individual drivers
2. In Cards.cpp Added stream overload
3. In map.cpp and .h fixed constructor and added copy constructors and destructors to Territory and Contenent classes
4. In player class fixed the addcardToHandOfCards function and fixed some of the constructor
5. Also in the player class changed playerID from const char* to string (therefore I also had to change this in the DriverSelector)
6. In the MapLoader class added a parameter constructor
Reasons:

1. due to the driverSelector we don't need the individual drivers anymore
2. Was missing
3. was missing or incomplete
4. was missing
5. as const char needs to be initialized when declared
6. So I do not need to run the functions set FileName, storeContinents, storeTerritories, and storeTerritoriesWithBorders it will do it automaticly
To Do:

Map.cpp needs setters for the Territory and Contenent Classes, line 31 needs someone who understands Lists to delete the list (maybe have to itterate accross the list I don't know) and Lines 52 and 69 needs someone who understands lists to assign the values from the original to the copyer, and line 300 needs to assign the default value to the destination (I am not sure if it is 0 or -1 or even if it makes a diffrence as I don't know if destination of the Territory is used at all).
Player.cpp the copy constructors and equals operator currently copy the same value need a tempThing =new type(*orig.thing) and then needs the tempThing to be pushed. Also functions such as addMyTerritory currently only prints that it did something does not actually do it.
Also in Player.cpp need to fix the destructor.
In Cards.cpp change from const string to enum (I am too lazy to do so as it was just sugested not mandatory).
Questions

1. Map.cpp line 330 why does it used two ways of checking if destination>=1 maybe it was supposed to check anther variable?